Unbreak the addCategoryToWatchlist.php maintenance script
WatchedItem::fromUserTitle() and WatchedItem::IGNORE_USER_RIGHTS were deprecated in MW 1.27 and removed in 1.31.
It seems that you don't need to use the WatchedItem to achieve the desired effect.

Lightly tested on my 1.39 box but should hopefully work on 1.35+ since that's what this extension supports, as per the extension.json file.

Change-Id: I6885171378153fd443e4f21c6c6267e1df34832a